Refactor: Move content generation handlers to a separate module
This commit refactors `App.jsx` by extracting five large handler functions into a new utility module at `src/utils/generationHandlers.js`.

The following functions were moved:
- `handleGenerateCampaignContent`
- `handleGenerateImage`
- `handleGenerateFormattedContent`
- `handleGenerateFollowupPosts`
- `handleGenerateIAContent`

The new module, `generationHandlers.js`, now contains the core logic for these operations as pure functions. `App.jsx` has been updated to import these functions and use them within smaller orchestrator handlers that are responsible for managing state and side effects.

This change improves code organization, separates concerns, and makes the `App.jsx` component significantly cleaner and easier to maintain.
