+++ src/App.jsx
@@ -99,6 +99,13 @@ import './App.css';
 import LoadingDialog from './components/LoadingDialog';
 import TextEditorDialog from './components/TextEditorDialog';
 import Campaign from './components/Campaign';
+import {
+  generateCampaignContent,
+  generateCampaignImage,
+  generateFormattedContent,
+  generateFollowupPosts,
+  generateIAContent,
+} from './utils/generationHandlers.js';
 
 // Temas atualizados com gradientes e cores modernas
 const lightTheme = createTheme({
@@ -1452,84 +1459,28 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
   }, [setCsvData]);
 
   const handleGenerateCampaignContent = async (regenerate = false) => {
-    if (!regenerate) {
-      setIsGeneratingCampaign(true);
-    } else {
-      // Se for apenas regeneração de texto, usar um estado de loading diferente se desejar
-      // Por enquanto, vamos usar o mesmo.
-      setIsGeneratingCampaign(true);
-    }
-
-    const apiKey = getGeminiApiKey();
-
-    if (!apiKey) {
-      alert('Por favor, configure sua chave de API Gemini primeiro.');
-      setIsGeneratingCampaign(false);
-      return;
-    }
-
-    const { persona, autor, instrucoes, formato } = getCampaignPrompt();
-
-    const promptCompleto = `
-      Persona: ${stripHtml(persona)}
-      Autor: ${stripHtml(autor)}
-      Formato: ${stripHtml(formato)}
-      Problema: ${stripHtml(problema)}
-      Solução: ${stripHtml(solucao)}
-      ${stripHtml(instrucoes)}
-    `;
-
-    const finalPrompt = `${promptCompleto}\n\nGere uma resposta JSON com os seguintes campos: "titulo" (string), "conteudo" (string), "cta" (string), e "hashtags" (string, separadas por vírgula). A resposta deve ser apenas o JSON.`;
-console.log(finalPrompt)
+    setIsGeneratingCampaign(true);
     try {
-      const response = await callGeminiApi(finalPrompt, apiKey);
-      console.log("Resposta da IA (Campanha):", response);
-
-      const jsonMatch = response.match(/```json\s*([\s\S]+?)\s*```/);
-      let parsedContent;
-
-      if (jsonMatch && jsonMatch[1]) {
-        parsedContent = JSON.parse(jsonMatch[1]);
-      } else {
-        parsedContent = JSON.parse(response);
-      }
-
-      let hashtags = [];
-      if (Array.isArray(parsedContent.hashtags)) {
-        hashtags = parsedContent.hashtags;
-      } else if (typeof parsedContent.hashtags === 'string') {
-        hashtags = parsedContent.hashtags.split(',').map(h => h.trim());
-      }
-
-      const normalizedContent = {
-        titulo: parsedContent.titulo || parsedContent.title || '',
-        conteudo: parsedContent.conteudo || parsedContent.body || '',
-        cta: parsedContent.cta || '',
-        hashtags: hashtags,
-      };
-
+      const normalizedContent = await generateCampaignContent({ problema, solucao });
       setCampaignContent(normalizedContent);
+
       if (!regenerate) {
         setConteudoMedio('');
         setConteudoPequeno('');
         setConteudoFormatado('');
         setGeneratedImageUrl(null);
-      }
 
-      // Se não for apenas regeneração, dispara as outras gerações
-      if (!regenerate) {
         await Promise.all([
           handleGenerateImage(normalizedContent),
           handleGenerateSummary(1800, normalizedContent),
           handleGenerateSummary(130, normalizedContent),
           handleGenerateFormattedContent(normalizedContent),
-          handleGenerateFollowupPosts(normalizedContent)
+          handleGenerateFollowupPosts(normalizedContent),
         ]);
       }
-
     } catch (error) {
       console.error("Erro ao gerar conteúdo da campanha:", error);
-      alert("Ocorreu um erro ao gerar o conteúdo da campanha. Verifique o console para mais detalhes.");
+      toast.error(`Ocorreu um erro ao gerar o conteúdo da campanha: ${error.message}`);
       setCampaignContent(null);
     } finally {
       setIsGeneratingCampaign(false);
@@ -1538,43 +1489,17 @@ console.log(finalPrompt)
 
   const handleGenerateImage = async (content = campaignContent) => {
     if (!content) {
-      alert("Por favor, gere o conteúdo do texto primeiro.");
+      toast.error("Por favor, gere o conteúdo do texto primeiro.");
       return;
     }
     setIsGeneratingImage(true);
-    const apiKey = getGeminiApiKey();
-    if (!apiKey) {
-      alert('Por favor, configure sua chave de API Gemini primeiro.');
-      setIsGeneratingImage(false);
-      return;
-    }
-
-    const { persona, autor, colors } = getCampaignPrompt();
     try {
-      const colorPalettePrompt = colors && colors.length > 0
-        ? `A imagem deve usar predominantemente a seguinte paleta de cores: ${colors.join(', ')}.`
-        : '';
-
-      const imagePrompt = `
-        Persona: ${stripHtml(persona)}
-        Autor: ${stripHtml(autor)}
-        Resumo do Conteúdo: ${stripHtml(content.titulo)}. ${stripHtml(content.conteudo)}
-        Razão de Aspecto: ${aspectRatio}
-        ${colorPalettePrompt}
-        ATENÇÃO: A imagem gerada não deve conter, sob NENHUMA CIRCUNSTÂNCIA, qualquer tipo de texto, escrita, letras, números ou palavras. A imagem deve ser puramente visual.
-      `;
-      const base64Image = await generateImage(imagePrompt, apiKey);
-
-      const imageUrl = `data:image/png;base64,${base64Image}`;
-
-      // A composição não é mais feita aqui.
-      // A imagem gerada pela IA é definida como a imagem de fundo e a imagem da campanha.
+      const imageUrl = await generateCampaignImage({ content, aspectRatio });
       setGeneratedImageUrl(imageUrl);
       updateImageAndPalette(imageUrl);
-
     } catch (imageError) {
       console.error("Erro ao gerar imagem:", imageError);
-      alert("Ocorreu um erro ao gerar a imagem da campanha. Verifique o console para mais detalhes.");
+      toast.error(`Ocorreu um erro ao gerar a imagem da campanha: ${imageError.message}`);
       setGeneratedImageUrl(null);
     } finally {
       setIsGeneratingImage(false);
@@ -1616,43 +1541,16 @@ console.log(finalPrompt)
 
   const handleGenerateFormattedContent = async (content = campaignContent) => {
     if (!content?.conteudo) {
-      alert("Por favor, gere o conteúdo principal primeiro.");
+      toast.error("Por favor, gere o conteúdo principal primeiro.");
       return;
     }
-
     setIsGeneratingConteudoFormatado(true);
-
-    const apiKey = getGeminiApiKey();
-    if (!apiKey) {
-      alert('Por favor, configure sua chave de API Gemini primeiro.');
-      setIsGeneratingConteudoFormatado(false);
-      return;
-    }
-
     try {
-      const prompt = `
-        Com o objetivo de gerar um post de blog no WordPress corporativo, Formatar o texto a seguir observando o padrão com HTML.
-        Considere que o conteúdo gerado já estará embutido em uma página no contexto de seu BODY.
-        Elabore o HTML para melhor estruturar o texto, facilitar a leitura, hierarquizar a informação conforme a importância.
-        O primeiro nível de Header que deve ser utilizado é o H3, já há H1 e H2 no contexto no qual o texto produzido se insere.
-        Elabore um resumo com os três pontos chave no texto de entrada e apresente o resumo com caixas de destaque logo no início.
-        ATENÇÃO aos campos que requeiram escape como aspas. Adicionalmente, o uso de &quot; é válido em HTML mas causa problemas em JSON. Atenção para evitar quebras de linha no conteúdo HTML e caracteres especiais não escapados.
-        Segue o texto:
-
-        Título: ${stripHtml(content.titulo)}
-        Conteúdo: ${stripHtml(content.conteudo)}
-        CTA: ${stripHtml(content.cta)}
-      `;
-
-      const rawContent = await callGeminiApi(prompt, apiKey);
-      // Remove markdown code block delimiters if they exist
-      const match = rawContent.match(/^`{3}(?:html)?\s*([\s\S]+?)\s*`{3}$/);
-      const finalContent = match && match[1] ? match[1].trim() : rawContent.trim();
+      const finalContent = await generateFormattedContent({ content });
       setConteudoFormatado(finalContent);
-
     } catch (error) {
       console.error(`Erro ao gerar conteúdo formatado:`, error);
-      alert(`Ocorreu um erro ao gerar o conteúdo formatado. Verifique o console.`);
+      toast.error(`Ocorreu um erro ao gerar o conteúdo formatado: ${error.message}`);
     } finally {
       setIsGeneratingConteudoFormatado(false);
     }
@@ -1660,91 +1558,16 @@ console.log(finalPrompt)
 
   const handleGenerateFollowupPosts = async (content = campaignContent) => {
     if (!content?.conteudo) {
-      alert("Por favor, gere o conteúdo principal primeiro.");
+      toast.error("Por favor, gere o conteúdo principal primeiro.");
       return;
     }
-
     setIsGeneratingFollowup(true);
-    const apiKey = getGeminiApiKey();
-    if (!apiKey) {
-      alert('Por favor, configure sua chave de API Gemini primeiro.');
-      setIsGeneratingFollowup(false);
-      return;
-    }
-
-    const { persona } = getCampaignPrompt();
-
     try {
-      const prompt = `
-        Você é um especialista em marketing de conteúdo e copywriting para líderes técnicos. Sua tarefa é criar ${followupPostsQuantity} posts "isca" baseados no conteúdo principal fornecido.
-
-        CONTEXTO:
-        O conteúdo principal aborda: [${stripHtml(content.titulo)} - ${stripHtml(content.conteudo)}]
-
-        PERSONAS-ALVO:
-        - ${stripHtml(persona)}
-
-        DIRETRIZES PARA OS POSTS:
-
-        1. Ganchos Psicológicos: Use gatilhos mentais como:
-           - Dor/Problema (rotatividade, custos, pressão)
-           - Curiosidade (estatísticas, casos reais)
-           - Urgência (mercado competitivo, riscos iminentes)
-           - Autoridade (experiência, casos de sucesso)
-           - Social Proof (situações reconhecíveis)
-
-        2. Estrutura de cada post:
-           - Hook inicial (pergunta provocativa ou estatística impactante)
-           - Desenvolvimento do problema/insight
-           - Call-to-action sutil direcionando para o conteúdo completo
-
-        3. Variação de Abordagens:
-           - Post 1: Foco na dor/problema
-           - Post 2: Estatística ou dado curioso
-           - Post 3: Caso real ou situação
-           - Post 4: Pergunta reflexiva
-           - Post 5: Insight contraintuitivo
-
-        ESPECIFICAÇÕES TÉCNICAS:
-        - Cada post deve ter entre 150-250 caracteres
-        - Tom profissional mas conversacional
-        - Inclua emojis estratégicos (máximo 2 por post)
-        - CTAs variados: "Leia mais", "Descubra como", "Saiba o que fazer"
-
-        FORMATO DE RESPOSTA:
-        Retorne um array JSON com a seguinte estrutura:
-
-        \`\`\`json
-        [
-          {
-            "post_numero": 1,
-            "tipo_gancho": "dor/problema",
-            "conteudo": "Texto do post aqui...",
-            "cta": "Call-to-action específico",
-            "hashtags_sugeridas": ["#liderancatecnica", "#gestaoequipes"]
-          }
-        ]
-        \`\`\`
-
-        OBJETIVO:
-        Cada post deve despertar curiosidade e criar um gap de informação que só será preenchido ao ler o conteúdo principal completo.
-      `;
-
-      const response = await callGeminiApi(prompt, apiKey);
-      const jsonMatch = response.match(/```json\s*([\s\S]+?)\s*```/);
-      let parsedContent;
-
-      if (jsonMatch && jsonMatch[1]) {
-        parsedContent = JSON.parse(jsonMatch[1]);
-      } else {
-        parsedContent = JSON.parse(response);
-      }
-
-      setFollowupPosts(parsedContent);
-
+      const posts = await generateFollowupPosts({ content, followupPostsQuantity });
+      setFollowupPosts(posts);
     } catch (error) {
       console.error(`Erro ao gerar posts de follow-up:`, error);
-      alert(`Ocorreu um erro ao gerar os posts de follow-up. Verifique o console.`);
+      toast.error(`Ocorreu um erro ao gerar os posts de follow-up: ${error.message}`);
     } finally {
       setIsGeneratingFollowup(false);
     }
@@ -1834,100 +1657,9 @@ console.log(finalPrompt)
 
   const handleGenerateIAContent = async () => {
     setIsGenerating(true);
-
-    let apiKey;
-    let apiToCall;
-    let apiName = "Gemini"; // Defaulting to Gemini
-
-    apiKey = getGeminiApiKey();
-    apiToCall = callGeminiApi;
-
-    if (!apiKey) {
-      alert(`Por favor, configure sua chave da API Gemini primeiro.\nVocê pode fazer isso no menu "Mais ações" (ícone de três pontos) no cabeçalho.`);
-      setIsGenerating(false);
-      return;
-    }
-
-    if (!promptText.trim()) {
-      alert('Por favor, forneça um texto descritivo para o prompt.');
-      setIsGenerating(false);
-      return;
-    }
-
-    if (promptNumRecords <= 0) {
-      alert('A quantidade de registros a gerar deve ser maior que zero.');
-      setIsGenerating(false);
-      return;
-    }
-
-    const finalPrompt = `A partir do TEXTO BASE fornecido abaixo, gere conteúdo para um carrossel de Instagram com ${promptNumRecords} elementos.
-
-TEXTO BASE:
-${stripHtml(promptText)}
-
-INSTRUÇÕES DE FORMATAÇÃO DA SAÍDA (MUITO IMPORTANTE):
-A SUA RESPOSTA DEVE CONTER *APENAS E SOMENTE* UM BLOCO DE TEXTO FORMATADO COMO CSV, SEM NENHUM TEXTO ADICIONAL ANTES OU DEPOIS DO BLOCO CSV.
-O BLOCO CSV DEVE SER DELIMITADO EXATAMENTE POR TRÊS CRASE SEGUIDAS E A PALAVRA "csv" (\`\`\`csv) NO INÍCIO, E TRÊS CRASE SEGUIDAS (\`\`\`) NO FINAL.
-DENTRO DO BLOCO CSV:
-- A primeira linha DEVE SER o cabeçalho: Titulo;Texto Principal;Ponte para o Próximo
-- As linhas subsequentes DEVERÃO ser os dados de cada elemento, com os campos separados por PONTO E VÍRGULA (;).
-- NÃO inclua números de elemento ou qualquer outra coluna além de "Titulo", "Texto Principal", e "Ponte para o Próximo".
-- NÃO inclua explicações, introduções, ou qualquer texto fora do bloco \`\`\`csv ... \`\`\`.
-
-REQUISITOS PARA O CONTEÚDO DE CADA ELEMENTO (LINHA DO CSV):
-1. **Titulo** (Coluna 1):
-   - Máximo de 4 palavras.
-   - Precisa ser curto e impactante.
-   - Exemplo: "Segredo Revelado"
-2. **Texto Principal** (Coluna 2):
-   - Entre 120 e 180 caracteres.
-   - Adaptado do TEXTO BASE, com linguagem conversacional e direta.
-   - Deve conter 1 pergunta retórica para engajamento.
-   - Exemplo: "Sabia que 80% dos negócios falham nisso? Descubra como evitar esse erro..."
-3. **Ponte para o Próximo** (Coluna 3):
-   - Máximo de 40 caracteres.
-   - Criar curiosidade para o próximo elemento.
-   - Usar fórmula: Emoji + Chamada + Dica do próximo.
-   - No último elemento, substitua por uma Chamada para Ação (CTA) final.
-   - Exemplos:
-     → "Próximo: O passo que muda tudo!"
-     → "Siga para o segredo nº3 👇"
-
-ESTRUTURA NARRATIVA SUGERIDA:
-- Elemento 1: Dado impactante ou pergunta instigante extraída do início do TEXTO BASE.
-- Elementos intermediários: Desenvolver os pontos principais do TEXTO BASE.
-- Último Elemento: CTA claro ou resumo conclusivo.
-
-TOM DE VOZ:
-- Empático e motivacional (use "você" e "vamos").
-- Urgência controlada ("Agora você pode...").
-- Toque de storytelling.
-
-Exemplo de como o BLOCO CSV deve se parecer na sua resposta (não inclua este exemplo na sua resposta final, apenas o bloco gerado):
-\`\`\`csv
-Titulo;Texto Principal;Ponte para o Próximo
-✨ Grande Novidade;Descubra algo incrível que vai mudar seu dia! Você está pronto para a surpresa?;➡️ Veja o próximo!
-🎉 Outra Dica;Continuando nossa jornada com mais um segredo. Já se perguntou como isso é possível?;CTA Final Aqui!
-\`\`\`
-Lembre-se: Sua resposta final deve conter APENAS o bloco \`\`\`csv ... \`\`\` com os dados.`;
-
-    console.log("Prompt para Gemini/DeepSeek:", finalPrompt); // Log atualizado para ser genérico
-    console.log("Número de Registros para Gerar:", promptNumRecords);
-
-    // console.log("Prompt para DeepSeek:", finalPrompt); // Manter para depuração se necessário
-    // console.log("Número de Registros para Gerar:", promptNumRecords);
-
     try {
-      let iaResponseText = "";
-      if (apiToCall) { // Verifica se apiToCall está definida
-        iaResponseText = await apiToCall(finalPrompt, apiKey);
-        console.log(`Resposta da API ${apiName} (bruta):`, iaResponseText);
-        console.log("Resposta da IA (Conteúdo):", iaResponseText);
-      } else {
-        throw new Error("Nenhuma função de API válida foi selecionada.");
-      }
-
-      const parsedResult = parseIaResponseToCsvData(iaResponseText, promptNumRecords);
+      const iaResponseText = await generateIAContent({ promptText, promptNumRecords });
+      const parsedResult = parseIaResponseToCsvData(iaResponseText); // This function stays in App.jsx
 
       if (parsedResult && parsedResult.data && parsedResult.data.length > 0) {
         setCsvData(parsedResult.data);
@@ -1955,13 +1687,12 @@ Lembre-se: Sua resposta final deve conter APENAS o bloco \`\`\`csv ... \`\`\` co
 
         setActiveStep(2); // Avança para Edição de Dados
       } else {
-        alert('Não foi possível processar a resposta da IA para o formato de tabela. Verifique o console para a resposta bruta da IA e a saída do parser.');
-        console.log(`[App] Falha no parsing ou dados vazios. Resposta da API ${apiName}:`, iaResponseText, "Resultado do Parser:", parsedResult);
+        toast.error('Não foi possível processar a resposta da IA para o formato de tabela.');
+        console.log(`[App] Falha no parsing ou dados vazios. Resposta da API:`, iaResponseText, "Resultado do Parser:", parsedResult);
       }
-
     } catch (error) {
-      console.error(`Erro ao chamar ou processar API ${apiName}:`, error);
-      alert(`Erro ao gerar conteúdo com IA via ${apiName}: ${error.message}`);
+      console.error(`Erro ao gerar conteúdo com IA:`, error);
+      toast.error(`Erro ao gerar conteúdo com IA: ${error.message}`);
     } finally {
       setIsGenerating(false);
     }
